test_autoVedio/requirements.txt
streamlit>=1.34,<2
langgraph>=0.2,<1
langchain>=0.2,<1
langchain-core>=0.2,<1
langchain-community>=0.2,<1
tavily-python>=0.3,<1
dashscope>=1.20,<2
openai-whisper>=20231117
moviepy>=2.0,<3
numpy>=1.26,<3
requests>=2.31,<3
imageio-ffmpeg>=0.4,<1
